correction deleteTransaction

+++ backend/controllers/transactionsController.js
@@ -164,13 +164,29 @@ export const updateTransaction = async (req, res) => {
 
 export const deleteTransaction = async (req, res) => {
   const { id } = req.params;
-  console.log(id);
   try {
     if (!id) {
       return res.status(404).json({ error: "Transaction not found" });
     }
+
+    // Trouver la transaction par son ID
+    const transaction = await Transaction.findById(id);
+
+    if (!transaction) {
+      return res.status(404).json({ error: "Transaction not found" });
+    }
+
+    // Supprimer la transaction de la base de données
     await Transaction.findByIdAndDelete(id);
-    res.status(200).json({ success: "transaction deleted successfully" });
+
+    // Mettre à jour le coin pour retirer l'ID de la transaction
+    await Coin.findByIdAndUpdate(transaction.coin, {
+      $pull: { transactions: id },
+    });
+
+    res
+      .status(200)
+      .json({ success: "Transaction deleted successfully and coin updated" });
   } catch (error) {
     res.status(500).json({ error: error.message });
   }
